backend/requirements.txt
# fastapi
# uvicorn[standard]
# sqlmodel
# python-dotenv
# google-generativeai
# pgvector
# langchain
# pydantic 
# psycopg2-binary
# numpy
# slowapi

fastapi
uvicorn
httpx
python-dotenv
google-generativeai